% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnwidget.R
\name{bnwidget}
\alias{bnwidget}
\title{Bayesian Network visuallization.}
\usage{
bnwidget(nodes, links, cpds, linkDistance = 150, opacity = 0.75,
  charge = -50, width = NULL, height = NULL)
}
\description{
bnwidget is an htmlwidget for Bayesian Network visualization
}
